app/requirements.txt
flake8==3.8.4
motor==2.3.0
pyflakes==2.2.0
pymongo==3.11.1
dnspython==2.0.0
email-validator==1.1.2
pydantic==1.7.3
python-multipart==0.0.5
python-jose==3.3.0
bcrypt==3.2.0